updated cj's scripts, fixed a few typos

+++ nnscripts/remove_parts_without_releases.php
@@ -1,87 +1,76 @@
 <?php
 /**
- * This parts which do not have a binary matched. 
- * 
- * The location of the script needs to be "misc/custom" or the
- * "misc/testing" directory. if used from another location,
- * change lines 34 to 36 to require the correct files.
+ * This script removes:
+ *  - parts which do not have a binary matched
+ *  - parts where the binary matched is missing
+ *  - binaries which do not have a part matched
  *
  * @author    NN Scripts
  * @license   http://opensource.org/licenses/MIT MIT License
  * @copyright (c) 2013 - NN Scripts
  *
  * Changelog:
- * 0.2 - Change retrieval of parts in groups of 10000
+ * 0.4 - Added remove of orphan binary records
+ * 
+ * 0.3 - Added settings.ini
+ *       Start using nnscript library
+ * 
+ * 0.2 - Change retrieval of parts in7groups of 10000
+ * 
  * 0.1 - Initial version
  */
-
 //----------------------------------------------------------------------
-// Settings
-
-// Display settings
-define('DISPLAY', true);
-
-// Time limit on releases to remove in hours
-// example: 24 for 1 day old releases (based on add date)
-// false or 0 to disable
-define('LIMIT', 36);
-
-// The number of parts that should be grouped when removing
-define('QUERYLIMIT', 10000);
-//----------------------------------------------------------------------
-
 // Load the application
-require(dirname(__FILE__)."/../bin/config.php");
-require_once(WWW_DIR."/lib/framework/db.php");
-require_once('nnscripts.php');
+require_once(dirname(__FILE__)."/../bin/config.php");
+
+// nnscripts includes
+require_once("lib/nnscripts.php");
 
 
 /**
  * Remove parts which are not linked to a binary
  */
-class removeParts
+class remove_parts_without_releases extends NNScripts
 {
     /**
-     * NNScripts class
-     * @var NNScripts
+     * The script name
+     * @var string
      */
-    private $nnscripts;     
+    protected $scriptName = 'Remove parts which are not linked to a binary';
     
     /**
-     * The database instance
-     * @db DB
+     * The script version
+     * @var string
      */
-    private $db;
-
+    protected $scriptVersion = '0.4';
+    
     /**
-     * The current date and time
-     * @var DateTime
+     * Allowed settings
+     * @var array
      */
-    private $now;
+    protected $allowedSettings = array('display', 'limit', 'querylimit');
 
-
-
+    
     /**
-     * Constructor
-     * 
-     * @param NNScripts $nnscripts
-     * @param DB $db
+     * The constructor
+     *
      */
-    public function __construct( NNScripts $nnscripts, DB $db )
+    public function __construct()
     {
-        // Set the NNScripts variable
-        $this->nnscripts = $nnscripts;
+        // Call the parent constructor
+        parent::__construct();
         
-        // Set the database variable
-        $this->db = $db;
+        // Set the commandline options
+        $options = array(
+            array( null, 'querylimit',  Ulrichsg\Getopt::REQUIRED_ARGUMENT, 'The query limit (default 10000)' ),
+        );
+        $this->setCliOptions( $options, array('display', 'limit', 'help') );
+        
+        // Show the header
+        $this->displayHeader();
 
-        // Add the "now" date
-        $this->now = new DateTime();
-        if( defined('LIMIT') && is_int(LIMIT) && 0 < LIMIT )
-        {
-            // substract the limit hours
-            $this->now->sub( new DateInterval('PT'. LIMIT .'H') );
-        }
+        // Show the settings
+        $this->displaySettings();
     }
     
     
@@ -97,9 +86,12 @@ class removeParts
         
         // Remove parts where the binary is missing
         $this->removePartsWithMissingBinary();
+        
+        // Remove binaries without parts
+        $this->removeBinariesWithoutParts();
     }
-
-
+    
+    
     /**
      * Remove parts without a binary linked
      * 
@@ -107,48 +99,21 @@ class removeParts
      */
     private function removePartsWithoutBinary()
     {
-        $this->nnscripts->display( "Removing old part records which are not linked to a binary: " );
+        $this->display( "Removing old part records which are not linked to a binary: " );
         
         // Build the sql
-        $sql = sprintf(
-            "DELETE FROM parts
-             WHERE parts.binaryID = 0
-             %s",
-            ( ( defined('LIMIT') && is_int(LIMIT) && 0 < LIMIT ) ? sprintf("AND parts.dateadded < '%s'", $this->now->format('Y-m-d H:i:s')) : '' )
-        );
-        $result = $this->db->queryDirect( $sql );
-        $this->nnscripts->display( ( $result ? "Done" : "Error" ) . PHP_EOL );
-    }
-
-
-    /**
-     * Get the total count of parts to remove
-     * 
-     * @return int
-     */
-    protected function getTotalPartsToRemove()
-    {
-        // Init
-        $ret = 0;
-
-        // Get the total to remove
-        $sql = sprintf("SELECT count(*) AS total
-                        FROM parts AS p
-                        LEFT JOIN binaries AS b ON (b.ID = p.binaryID)
-                        WHERE b.ID IS NULL
-                        %s",
-                        ( ( defined('LIMIT') && is_int(LIMIT) && 0 < LIMIT ) ? sprintf("AND p.dateadded < '%s'", $this->now->format('Y-m-d H:i:s')) : '' ) );
-        $result = $this->db->query( $sql );
-        if( is_array($result) && 1 === count($result) )
+        $sql = "DELETE FROM parts
+                WHERE binaryID = 0";
+        if( is_numeric( $this->settings['limit'] ) && 0 < $this->settings['limit'] )
         {
-            $ret = $result[0]['total'];
+            $sql .= sprintf( ' AND dateadded < "%s" - INTERVAL %s HOUR', $this->settings['now'], $this->settings['limit'] );
         }
-
-        // Return
-        return $ret;
+        $result = $this->db->queryDirect( $sql );
+        
+        $this->display( ( $result ? "Done" : "Error" ) . PHP_EOL );
     }
     
-
+    
     /**
      * Remove parts where the binary is missing
      * 
@@ -156,18 +121,17 @@ class removeParts
      */
     private function removePartsWithMissingBinary()
     {
-        // The query limit
-        $queryLimit = ( defined('QUERYLIMIT') && is_int(QUERYLIMIT) && 0 < QUERYLIMIT ? QUERYLIMIT : 10000 );
-        
         // Get the total parts to remove
         $total = $this->getTotalPartsToRemove();
         if( 0 < $total )
         {
             $line = sprintf( "Removing %s parts with missing binaries", number_format($total, 0, '.', '') );
-            if( $total > $queryLimit )
-                $line .= sprintf( ' (in groups of %d parts)', $queryLimit );
+            if( $total > $this->settings['querylimit'] )
+            {
+                $line .= sprintf( ' (in groups of %d parts)', $this->settings['querylimit'] );
+            }
             $line .= ': ';
-            $this->nnscripts->display( $line );
+            $this->display( $line );
             
             // Set the lastID
             $lastId = null;
@@ -186,8 +150,12 @@ class removeParts
                                 ORDER BY p.ID ASC
                                 LIMIT %d",
                     ( null !== $lastId ? sprintf('AND p.ID > %d', $lastId) : '' ),
-                    ( ( defined('LIMIT') && is_int(LIMIT) && 0 < LIMIT ) ? sprintf("AND p.dateadded < '%s'", $this->now->format('Y-m-d H:i:s')) : '' ),
-                    $queryLimit
+                    (
+                        ( is_numeric( $this->settings['limit'] ) && 0 < $this->settings['limit'] )
+                        ? sprintf( ' AND p.dateadded < "%s" - INTERVAL %s HOUR', $this->settings['now'], $this->settings['limit'] )
+                        : ''
+                    ),
+                    $this->settings['querylimit']
                 );
                 $result = $this->db->query( $sql );
                 if( is_array( $result ) && 0 < count( $result ) )
@@ -199,9 +167,9 @@ class removeParts
                     {
                         $ids[] = $lastId = $row['ID'];
                         $counter++;
-                        if( $queryLimit === $counter )
+                        if( $this->settings['querylimit'] === $counter )
                         {
-                            $this->nnscripts->display( "." );
+                            $this->display( "." );
                             $sql = sprintf( "DELETE FROM parts WHERE ID IN (%s)", implode(',', $ids) );
                             $this->db->queryDirect( $sql );
                             $ids = array();
@@ -213,7 +181,7 @@ class removeParts
                     unset( $result );
                     
                     // Lower the total
-                    $total -= $queryLimit;
+                    $total -= $this->settings['querylimit'];
                 }
                 else
                 {
@@ -223,43 +191,81 @@ class removeParts
             }
             
             // And remove the rest
-            if( 0 < count($ids) )
+            if( is_array($ids) && 0 < count($ids) )
             {
-                $this->nnscripts->display( "." );
+                $this->display( "." );
                 $sql = sprintf( "DELETE FROM parts WHERE ID IN (%s)", implode(',', $ids) );
                 $this->db->queryDirect( $sql );
             }
             
-            $this->nnscripts->display( PHP_EOL );
+            $this->display( PHP_EOL );
         }
         else
         {
-            $this->nnscripts->display( 'No parts with missing binaries to remove.'. PHP_EOL );
+            $this->display( 'No parts with missing binaries to remove.'. PHP_EOL );
         }
+    }
+    
+    
+    /**
+     * Get the total count of parts to remove
+     * 
+     * @return int
+     */
+    protected function getTotalPartsToRemove()
+    {
+        // Init
+        $ret = 0;
+
+        // Get the total to remove
+        $sql = "SELECT count(*) AS total
+                FROM parts AS p
+                LEFT JOIN binaries AS b ON (b.ID = p.binaryID)
+                WHERE b.ID IS NULL";
+        if( is_numeric( $this->settings['limit'] ) && 0 < $this->settings['limit'] )
+        {
+            $sql .= sprintf( ' AND p.dateadded < "%s" - INTERVAL %s HOUR', $this->settings['now'], $this->settings['limit'] );
+        }
+        
+        $result = $this->db->queryOneRow( $sql );
+        if( is_array($result) && array_key_exists('total', $result) )
+        {
+            $ret = (int)$result['total'];
+        }
+
+        // Return
+        return $ret;
+    }
+    
+    
+    /**
+     * Remove orhpan binary records
+     * 
+     * @return void
+     */
+    protected function removeBinariesWithoutParts()
+    {
+        $this->display( "Removing old binary records which do not have parts: " );
+        
+        $sql = "DELETE binaries FROM binaries
+                LEFT JOIN parts ON (parts.binaryID = binaries.ID)
+                WHERE parts.ID IS NULL";
+        if( is_numeric( $this->settings['limit'] ) && 0 < $this->settings['limit'] )
+        {
+            $sql .= sprintf( ' AND binaries.dateadded < "%s" - INTERVAL %s HOUR', $this->settings['now'], $this->settings['limit'] );
+        }
+        $result = $this->db->queryDirect( $sql );
+        
+        $this->display( ( $result ? "Done" : "Error" ) . PHP_EOL );
     }
 }
 
+// Main application
 try
 {
-    // Init
-    $scriptName    = 'Remove parts which are not linked to a binary';
-    $scriptVersion = '0.2';
-    
-    // Load the NNscript class
-    $nnscripts = new NNScripts( $scriptName, $scriptVersion );
-    
-    // Display the header
-    $nnscripts->displayHeader();
-
-    // Load the application part
-    $db = new DB;
-    if( !$db )
-        throw new Exception("Error loading database library");
-        
-    // Load the categoryReleases class
-    $cr = new removeParts( $nnscripts, $db );
-    $cr->cleanup();
+    // Load the remove_parts_without_releases class
+    $rpwr = new remove_parts_without_releases();
+    $rpwr->cleanup();
 } catch( Exception $e ) {
     echo $e->getMessage() . PHP_EOL;
 }
-
